rating c + funtion work and gui work
add TBV to show up
chance dubble to int on rating
add new delete scene
add add buten sceen to work
and more

+++ src/unwanted_mc/bll/BllManager.java
@@ -39,7 +39,7 @@ public class BllManager implements IBLL {
  
   
     @Override
-    public Movie addMovieToDB(String name, double rating, String filelink, String lastview) {
+    public Movie addMovieToDB(String name, int rating, String filelink, String lastview) {
         return dalManager.addMovieToDB(name, rating, filelink, lastview);
     }
         
@@ -63,7 +63,7 @@ public class BllManager implements IBLL {
 
     
     @Override
-    public Movie editMovie(String name, double rating, String filelink, String lastview) {
+    public Movie editMovie(String name, int rating, String filelink, String lastview) {
         return dalManager.editMovie(name, rating, filelink, lastview);
     }
