+++ src/unwanted_mc/dal/DalManager.java
@@ -33,7 +33,7 @@ public class DalManager implements IDAL {
     
     
     @Override
-    public Movie addMovieToDB(String name, double rating, String filelink, String lastview) {
+    public Movie addMovieToDB(String name, int rating, String filelink, String lastview) {
         return movieDBDao.addMovieToDB(name, rating, filelink, lastview);
     }
         
@@ -62,7 +62,7 @@ public class DalManager implements IDAL {
 
     
     @Override
-    public Movie editMovie(String name, double rating, String filelink, String lastview) {
+    public Movie editMovie(String name, int rating, String filelink, String lastview) {
         return  movieDBDao.editMovie(name, rating, filelink, lastview);
     }
